requirements.txt
appdirs==1.4.4
asgiref==3.4.1
black==21.7b0
click==8.0.1
colorama==0.4.4
Django==3.2.6
django-debug-toolbar==3.2.2
djangorestframework==3.12.4
mypy-extensions==0.4.3
pathspec==0.9.0
pytz==2021.1
regex==2021.8.21
six==1.16.0
sqlparse==0.4.1
tomli==1.2.1
